compiler: Move Stamp from ir/space to tools/abc
